+++ R/cat2cat_agg.R
@@ -41,7 +41,7 @@ cat2cat_agg <- function(data = list(
                         ), ...) {
   # methods most frequent, proportional
 
-  assert_that(
+  stopifnot(
     is.list(data) &&
       length(data) == 5 &&
       all(vapply(data, Negate(is.null), logical(1))) &&
@@ -55,9 +55,9 @@ cat2cat_agg <- function(data = list(
   d_old <- length(unique(data$old[[data$time_var]]))
   d_new <- length(unique(data$new[[data$time_var]]))
 
-  assert_that((d_old == 1) && (d_new == 1))
+  stopifnot((d_old == 1) && (d_new == 1))
 
-  assert_that(all(table(data$old[[data$cat_var]]) == 1) && all(table(data$new[[data$cat_var]]) == 1))
+  stopifnot(all(table(data$old[[data$cat_var]]) == 1) && all(table(data$new[[data$cat_var]]) == 1))
 
   t <- as.list(substitute(list(...))[-1])
 
@@ -72,10 +72,10 @@ cat2cat_agg <- function(data = list(
 
   col_df_new <- colnames(df_new)
 
-  assert_that(identical(col_df_old, col_df_new))
+  stopifnot(identical(col_df_old, col_df_new))
 
   for (i in trans_map) {
-    assert_that(((length(i[[2]]) == 1) || (length(i[[3]]) == 1)))
+    stopifnot(((length(i[[2]]) == 1) || (length(i[[3]]) == 1)))
 
     if (i$direction == "forward") {
       base <- df_new[!(df_new[, data$cat_var] %in% i[[3]]), ]